Rename aux to std since it's somekind of standard library.
